+++ lib/components/AutoComplete/AutoComplete.d.ts
@@ -1,0 +1,3 @@
+import React from 'react';
+import { IAutoCompleteProps } from './AutoComplete.models';
+export declare const AutoComplete: React.FC<IAutoCompleteProps>;