+++ lib/components/AutoComplete/AutoComplete.styles.d.ts
@@ -1,0 +1,14 @@
+export declare const autoCompleteStyles: (theme: string) => {
+    autoCompleteLabel: {
+        fontWeight: number;
+    };
+    autoCompleteError: {
+        outline: string;
+        outlineWidth: string;
+    };
+    errorText: {
+        color: string;
+        padding: string;
+        fontSize: string;
+    };
+};