+++ src/components/AutoComplete/AutoComplete.styles.ts
@@ -1,0 +1,14 @@
+export const autoCompleteStyles = (theme: string) => ({
+  autoCompleteLabel: {
+    fontWeight: 400
+  },
+  autoCompleteError: {
+    outline: '1px solid red',
+    outlineWidth: 'thin'
+  },
+  errorText: {
+    color: 'red',
+    padding: '0px 0px',
+    fontSize: '14px'
+  }
+});