+++ src/components/Wizard/Wizard.models.ts
@@ -71,4 +71,5 @@ export interface IWizardProps {
   stepActiveClass?: string;
   wizardLinksClass?: string;
   wizardContainerClass?: string;
+  wizardLinkTextClass?: string;
 }