+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './Accordion';
+export * from './AutoComplete';
 export * from './AppWrapper';
 export * from './AzureAD';
 export * from './ErrorBoundary';